Updated LoginView, LoginControl. Moves util files to utility package. Used envUtil inside EmailUtil.
