+++ src/.env.example
@@ -1,0 +1,4 @@
+HOST=smtp.gmail.com
+PORT=587
+USERNAME=ankanaghosh1111@gmail.com
+PASWORD=vjuc xorq kafn mvnm
