--- src/control/config.java
@@ -1,8 +1,0 @@
-package control;
-// Utility class to store configs
-public class config {
-    // Location of various database files
-    public static final String empDBlocation = "C:/Users/DELL/Desktop/EmpDataDB.csv";
-    public static final String empFileLocation = "C:/Users/DELL/Desktop/EmpData.db";
-    public static final String credentialsFileLocation = "C:/Users/DELL/Desktop/Creds.db";
-}
