+++ src/utility/credUtil.java
@@ -1,5 +1,7 @@
-package control;
+package utility;
 import java.io.File;
+import config;
+
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
